Time: 319 ms (100.00%), Space: 120.4 MB (75.00%) - LeetHub

+++ 2588-count-the-number-of-beautiful-subarrays/2588-count-the-number-of-beautiful-subarrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long beautifulSubarrays(vector<int>& nums) {
         long long count=0;
-        unordered_map<int,long long>mp;
+        unordered_map<int,int>mp;
         mp[0]=1;
         for(int i=0,sum=0,n=nums.size();i<n;i++)
         {
